+++ packages/core/src/mm/play/jp_layout.c
@@ -24,7 +24,7 @@ static void ComboPlay_SpawnLayoutDependantTorch(PlayState* play, float x, float 
     Actor_Spawn(&play->actorCtx, play, ACTOR_OBJ_SYOKUDAI, x, y, z, 0, 0, 0, params);
 }
 
-void ComboPlay_JpLayout(PlayState* play)
+static void ComboPlay_JpLayoutDekuPalace(PlayState* play)
 {
     if (!Config_Flag(CFG_MM_JP_LAYOUT_DEKU_PALACE) || play->sceneId != SCE_MM_DEKU_PALACE)
         return;
@@ -46,6 +46,17 @@ void ComboPlay_JpLayout(PlayState* play)
     case 0x02:
         ComboPlay_SpawnLayoutDependantTorch(play, -1040, 0, 656);
         ComboPlay_SpawnLayoutDependantTorch(play, -420, 0, 1295);
+        break;
+    }
+}
+
+void ComboPlay_JpLayout(PlayState* play)
+{
+    switch (play->sceneId)
+    {
+    case SCE_MM_DEKU_PALACE:
+        if (Config_Flag(CFG_MM_JP_LAYOUT_DEKU_PALACE))
+            ComboPlay_JpLayoutDekuPalace(play);
         break;
     }
 }